Added a setup function that can be run in the background, but anything that need to use the renderer needs to be on the main thread(at least for now) so should happen in the constructor.
Also set the "nextScreen" pointer in the constructor of the base screen rather than in the header to stop this from complaining when I commit this file and made it print out the delta time in the screen to make it stop saying that delta time is not used... I know, not yet, but soon. *roll eyes*

Lastly the LoadScreen now handles input(you can quit while loading.)

I can't figure out why it says the load screen is never used, so I will kick that can further down the road.