Added some try-catch and print statements for debugging.
I found the problem that was causing the "race" condition(or so I think...)

It wasn't actually a race condition, but I don't know how it actually worked in the first place. I wrote an improper CheckSetup for the GameScreen, I left it from an earlier version and never updated it to using the map. I don't know how the texture wasn't a nullptr some runs.

Well I haven't noticed crashing now, so I think this is better.